Add div around hero image
This simplifies the sizing the hero image by allowing to set its width
to 100% of its parent container - the new div. The new div is 50% of
`.hero`. Though, I'm not sure what the difference is between this and
simply omitting the div and setting image width to 50%. They produce
different outputs for some reason.